requirements.txt
kafka-python == 2.02